.history/requirements_20240502023803.txt
youtube_transcript_api
streamlit
google-generativeai
google-cloud
python-dotenv
pathlib

-e.